# initial sorted search implementation