1. Refactored feature search for the usage of permutations and lineat search. 

Benchmark increase: 
- decision tree (100x100 matrix, 3 classes): from 7.5ms to 3.6ms (median time)
- random forest (300x100 matrix, 3 classes, 50 trees): from 317ms to 90ms (median time)

2. Removed obsolete function `test_split`
3. Refactored tests.